refactor: SocialMedia to shared
