+++ src/app/components/index.ts
@@ -8,5 +8,4 @@ export { OptionItem, type OptionItemProps } from './option-item';
 export { Places } from './places';
 export { PrincipleCard, type PrincipleCardProps } from './principle-card';
 export { SchoolMenu } from './school-menu';
-export { SocialMedia, type SocialMediaProps } from './social-media';
 export { TextWithLink } from './text-with-link';
