+++ src/features/home/community/community.tsx
@@ -1,8 +1,7 @@
-import { SocialMedia, type SocialMediaProps } from '@/app/components';
-
 import image from '@/assets/welcome.webp';
 import Image from '@/features/image';
 import { FacebookIcon, InstagramIcon, LinkedInIcon, TelegramIcon, YouTubeIcon } from '@/icons';
+import { SocialMedia, SocialMediaProps } from '@/shared/ui/social-media';
 import { Subtitle } from '@/shared/ui/subtitle';
 import { Title, TitleType } from '@/shared/ui/title';
