+++ src/features/home/pictures/pictures.tsx
@@ -1,6 +1,5 @@
 // requires a loader
 import { Carousel } from 'react-responsive-carousel';
-import { SocialMedia, SocialMediaProps } from '@/app/components';
 import photo1 from '@/assets/photo-1.webp';
 import photo2 from '@/assets/photo-2.webp';
 import photo3 from '@/assets/photo-3.webp';
@@ -13,6 +12,7 @@ import photo9 from '@/assets/photo-9.webp';
 import Image from '@/features/image';
 import { FacebookIcon, InstagramIcon } from '@/icons';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { SocialMedia, SocialMediaProps } from '@/shared/ui/social-media';
 import { Title } from '@/shared/ui/title';
 
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
