+++ src/shared/ui/social-media/social-media.scss
