+++ src/shared/ui/social-media/social-media.test.tsx
